refactoring: modules clean up
