+++ tg-stonks-bot/src/tg_stonks/database/entity_models.py
