+++ tg-stonks-bot/src/tg_stonks/database/errors.py
