+++ tg-stonks-bot/src/tg_stonks/database/helpers.py
@@ -3,9 +3,9 @@ from typing import Callable, Any
 
 from returns.result import Success, Result, safe
 
-from .entity_models import UserEntity, InstrumentEntity, TrackingEntity
-from tg_stonks_bot.database.user_settings import DataProviderConfig, UserSettings
-from .protocols import IDatabase
+from tg_stonks.database.entity_models import UserEntity, InstrumentEntity, TrackingEntity
+from tg_stonks.database.user_settings import DataProviderConfig, UserSettings
+from tg_stonks.database.protocols import IDatabase
 
 
 def to_user(user_obj: dict) -> UserEntity:
