+++ tg-stonks-bot/src/tg_stonks/database/protocols.py
