+++ tg-stonks-bot/src/tg_stonks/database/user_settings.py
@@ -2,7 +2,7 @@ from typing import Optional
 
 from pydantic import BaseModel, validator
 
-from ..data_providers.provider_type import ProviderT
+from tg_stonks.providers.provider_type import ProviderT
 
 
 class DataProviderConfig(BaseModel):
