+++ tg-stonks-bot/src/tg_stonks/impl/alpha_vantage_provider.py
@@ -4,20 +4,18 @@ from alpha_vantage.async_support.cryptocurrencies import CryptoCurrencies as Cry
 from alpha_vantage.async_support.foreignexchange import ForeignExchange as ForeignExchangeAsync
 from alpha_vantage.async_support.timeseries import TimeSeries as TimeSeriesAsync
 
-from .data_providers.api_client_middleware import ApiClientMiddleware
-
-from .data_providers.protocols import (
-    IDataProviderStockMarket,
-    IDataProviderCurrencyEx,
-    IDataProviderCryptoEx
-)
-
-from .data_providers.provider_type import ProviderT
-from tg_stonks_bot.alpha_vantage_models import (
+from tg_stonks.impl.alpha_vantage_models import (
     StockMarketInstrumentAV,
     SearchQueryResAV,
     ExchangePairAV
 )
+from tg_stonks.providers.api_client_middleware import ApiClientMiddleware
+from tg_stonks.providers.protocols import (
+    IDataProviderStockMarket,
+    IDataProviderCurrencyEx,
+    IDataProviderCryptoEx
+)
+from tg_stonks.providers.provider_type import ProviderT
 
 
 @final
