+++ tg-stonks-bot/src/tg_stonks/providers/models.py
