+++ tg-stonks-bot/src/tg_stonks/providers/protocols.py
@@ -1,8 +1,8 @@
 from abc import abstractmethod
 from typing import Protocol, runtime_checkable
 
-from .provider_type import ProviderT
-from tg_stonks_bot.data_providers.models import ExchangePair, StockMarketInstrument, SearchQueryRes
+from tg_stonks.providers.provider_type import ProviderT
+from tg_stonks.providers.models import ExchangePair, StockMarketInstrument, SearchQueryRes
 
 
 # Base protocol for data providers
