+++ tg-stonks-bot/src/tg_stonks/providers/provider_type.py
