+++ tg-stonks-bot/src/tg_stonks/upd_listener/protocol.py
